serial: sh-sci: Fix scatterlist mapping leak
The mapped scatterlist is never unmapped. This leaks quite some
mappings, as the mapping is done in uart_ops.startup(), i.e. every time
the device is opened. Unmap the scatterlist on device close.

Note that we have to restore the original DMA entry length and DMA
address, as it has been modified in work_fn_tx(). Else warnings are
printed if CONFIG_DMA_API_DEBUG=y:

    WARNING: CPU: 0 PID: 20 at lib/dma-debug.c:1103 check_unmap+0x24c/0x85c()
    rcar-dmac e6700000.dma-controller: DMA-API: device driver frees DMA memory with different size [device address=0x000000006e15f000] [map size=4096 bytes] [unmap size=3 bytes]

and:

    WARNING: CPU: 0 PID: 1364 at lib/dma-debug.c:1093 check_unmap+0x178/0x85c()
    rcar-dmac e6700000.dma-controller: DMA-API: device driver tries to free DMA memory it has not allocated [device address=0x000000006e16b006] [size=4096 bytes]

TODO: Simplify by switching to dma_map_single(), as the transmit
scatterlist contains one single entry, which is reused and modified.

Signed-off-by: Geert Uytterhoeven <geert+renesas@glider.be>
---
v2:
  - Add TODO