serial: sh-sci: Add (H)SCIF DMA support
When using receive DMA on SCIF and HSCIF, an interrupt is triggered only
after the whole DMA request has been transferred. If less data is
received, it will stay stuck until more data arrives.

To make sure we are notified as soon as any data has arrived, an
additional one-byte DMA buffer is used. Once the first byte of data is
received, the traditional larger DMA buffers are used, and the receive
DMA timeout system will take care of receiving partial data.

This was inspired by a patch in the BSP by Koji Matsuoka
<koji.matsuoka.xm@renesas.com>.

Signed-off-by: Geert Uytterhoeven <geert+renesas@glider.be>